Web_Scrappers/requirements.txt
selenium==3.141.0
webdriver-manager==3.4.2
urllib3==1.26.7
pytesseract==0.3.8
Pillow==8.4.0
opencv-python==4.5.3.56
easyocr==1.3.2
numpy==1.21.2
pywin32==301
